Make shader compilation optional and allow using raw spirv
