+++ rendy/Cargo.toml
@@ -18,6 +18,7 @@ metal = ["rendy-factory/metal", "rendy-graph/metal", "rendy-wsi/metal", "gfx-bac
 vulkan = ["rendy-factory/vulkan", "rendy-graph/vulkan", "rendy-wsi/vulkan", "gfx-backend-vulkan"]
 serde-1 = ["gfx-hal/serde", "rendy-factory/serde-1", "rendy-mesh/serde-1", "rendy-texture/serde-1"]
 no-slow-safety-checks = ["rendy-factory/no-slow-safety-checks", "rendy-wsi/no-slow-safety-checks", "rendy-util/no-slow-safety-checks"]
+shader-compiler = ["rendy-shader/shader-compiler"]
 
 command = ["rendy-command"]
 factory = ["rendy-factory", "command", "resource", "wsi"]
@@ -31,7 +32,7 @@ texture = ["rendy-texture", "factory", "util"]
 util = ["rendy-util"]
 wsi = ["rendy-wsi"]
 
-base = ["command", "factory", "frame", "graph", "memory", "mesh", "shader", "resource", "texture", "util", "wsi"]
+base = ["shader-compiler", "command", "factory", "frame", "graph", "memory", "mesh", "shader", "resource", "texture", "util", "wsi"]
 default = ["base"]
 
 mesh-obj = ["mesh", "rendy-mesh/obj"]
