+++ shader/Cargo.toml
@@ -14,4 +14,7 @@ description = "Rendy's shader compilation tool"
 failure = "0.1"
 gfx-hal = "0.1"
 rendy-factory = { version = "0.1", path = "../factory" }
-shaderc = "0.3"
+shaderc = { version = "0.3", optional = true }
+
+[features]
+shader-compiler = ["shaderc"]
